Add integration with Home Assistant MQTT autodiscovery
Home Assistant can automatically create sensor objects and attach data to
them if appropriate config information is provided by the sensor. This
implementation fires far too many config messages (it should do it on startup
and then subscribe to HA's online/offline messages to determine when to do
so again), but this works for now.